+++ event-sourcing/event-publisher/published-event-processor/src/main/java/uk/gov/justice/services/eventsourcing/publishedevent/PublishQueueException.java
@@ -1,0 +1,12 @@
+package uk.gov.justice.services.eventsourcing.publishedevent;
+
+public class PublishQueueException extends RuntimeException {
+
+    public PublishQueueException(final String message) {
+        super(message);
+    }
+
+    public PublishQueueException(final String message, final Throwable cause) {
+        super(message, cause);
+    }
+}